requirements.txt
Flask
flask-cors
openai
requests
replicate
python-dotenv
aiohttp
reportlab
gunicorn
cloudinary
pymongo
flask-pymongo
flask-login
google-auth
google-auth-oauthlib
PyJWT==2.8.0
